Make concept representation explicit in schema
Previously, concepts were entities that had a special attribute set (`is_concept`). They were
expected to have `name` set, but this wasn't enforced outside of the standard helpers.
Adding a concept table enables the enforcement of name uniqueness for concepts at the database level
Adding an instance_of table to replace the "instance_of" relation enforces that the relationship must be between an entity and a concept
This reduces the power of API consumers to directly manipulate this attribute (just look at how much of Concept's implementation had to be
pushed directly into LTMC), but in the end this probably what we want for such a sensitive relation. We should consider the same move
for `is_a` at some point
Implemented additional queries for getting concept children/ancestors/instances recursively as SQL functions
Breaking API change: made non-recursive the default for removeInstances and getConcepts